# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kat's Beauty Bar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  active: boolean;
}

export interface ServiceInput {
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ServiceUpdate {
  name?: string;
  description?: string;
  price?: number;
  durationMinutes?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  active?: boolean;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  serviceId: number;
  serviceName: string;
  /** ISO 8601 datetime */
  startTime: string;
  /** ISO 8601 datetime */
  endTime: string;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface AppointmentInput {
  customerName: string;
  email: string;
  phone: string;
  serviceId: number;
  /** ISO 8601 datetime */
  startTime: string;
  /** @nullable */
  notes?: string | null;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface AppointmentUpdate {
  startTime?: string;
  status?: AppointmentUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface TimeSlot {
  /** HH:MM format */
  time: string;
  available: boolean;
  /**
     * Why unavailable (booked, closed, etc.)
     * @nullable
     */
  reason?: string | null;
}

export interface DayHours {
  open: boolean;
  /** @nullable */
  openTime?: string | null;
  /** @nullable */
  closeTime?: string | null;
}

export interface AvailabilityResponse {
  date: string;
  slots: TimeSlot[];
  businessHours?: DayHours;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
}

export interface LoginResult {
  token: string;
  admin: AdminUser;
}

export interface ServicePopularity {
  serviceName: string;
  bookingCount: number;
}

export interface DashboardStats {
  todayAppointments: number;
  upcomingAppointments: number;
  monthlyBookings: number;
  popularServices: ServicePopularity[];
}

export type BusinessSettingsOpeningHours = {[key: string]: DayHours};

export interface BusinessSettings {
  slotIntervalMinutes: number;
  ownerEmail: string;
  openingHours: BusinessSettingsOpeningHours;
}

export type BusinessSettingsUpdateOpeningHours = {[key: string]: DayHours};

export interface BusinessSettingsUpdate {
  slotIntervalMinutes?: number;
  ownerEmail?: string;
  openingHours?: BusinessSettingsUpdateOpeningHours;
}

export type ListAppointmentsParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
/**
 * Filter by status
 */
status?: string;
};

export type GetAvailabilityParams = {
/**
 * Date in YYYY-MM-DD format
 */
date: string;
/**
 * Service ID to check duration
 */
serviceId: number;
};