# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kat's Beauty Bar API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  startTime?: string;
  status?: AppointmentUpdateStatus;
  /** @nullable */
  notes?: string | null;
}