# lib/api-zod/src/generated/types/availabilityResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kat's Beauty Bar API
 * OpenAPI spec version: 0.1.0
 */
import type { DayHours } from './dayHours';
import type { TimeSlot } from './timeSlot';

export interface AvailabilityResponse {
  date: string;
  slots: TimeSlot[];
  businessHours?: DayHours;
}