# lib/api-zod/src/generated/types/businessSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kat's Beauty Bar API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessSettingsUpdateOpeningHours } from './businessSettingsUpdateOpeningHours';

export interface BusinessSettingsUpdate {
  slotIntervalMinutes?: number;
  ownerEmail?: string;
  openingHours?: BusinessSettingsUpdateOpeningHours;
}