# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kat's Beauty Bar API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  name?: string;
  description?: string;
  price?: number;
  durationMinutes?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  active?: boolean;
}